Ajuste para alterar o tipo de dado devido ao tipo de banco de dados mysql

+++ udemy/angular_spring_boot/helpdesk/src/main/java/com/aug/helpdesk/domain/Cliente.java
@@ -1,6 +1,5 @@
 package com.aug.helpdesk.domain;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,7 +24,7 @@ public class Cliente extends Pessoa {
         addPerfis(Perfil.CLIENTE);
     }
     
-    public Cliente(BigInteger id, String nome, String cpf, String email, String senha) {
+    public Cliente(Integer id, String nome, String cpf, String email, String senha) {
         super(id, nome, cpf, email, senha);
         addPerfis(Perfil.CLIENTE);
     }
